refactor: moved mixins and styled components to their own folder. moved drawer width parameter to constants folder
