+++ src/components/Header/Header.types.ts
@@ -1,0 +1,6 @@
+import { AppBarProps as MuiAppBarProps } from '@mui/material';
+
+export interface AppBarProps extends MuiAppBarProps {
+    discreet?: boolean;
+    open?: boolean;
+}
