+++ src/components/SidebarStatic/SidebarStatic.styles.tsx
@@ -1,0 +1,57 @@
+import { Drawer as MuiDrawer, styled, Theme, CSSObject } from '@mui/material';
+
+import { headerDrawerOpenWidth } from '../../constants/appConstants';
+
+const openedMixin = (theme: Theme): CSSObject => ({
+    overflowX: 'hidden',
+    transition: theme.transitions.create('width', {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.enteringScreen,
+    }),
+    width: headerDrawerOpenWidth,
+});
+
+const closedMixin = (theme: Theme): CSSObject => ({
+    overflowX: 'hidden',
+    transition: theme.transitions.create('width', {
+        easing: theme.transitions.easing.sharp,
+        duration: theme.transitions.duration.leavingScreen,
+    }),
+    width: `calc(${theme.spacing(7)} + 1px)`,
+    [theme.breakpoints.up('sm')]: {
+        width: `calc(${theme.spacing(8)} + 1px)`,
+    },
+});
+
+export const DrawerHeader = styled('div')(({ theme }) => ({
+    alignItems: 'center',
+    display: 'flex',
+    justifyContent: 'flex-end',
+    padding: theme.spacing(0, 1),
+    ...theme.mixins.toolbar,
+}));
+
+export const Drawer = styled(MuiDrawer, {
+    shouldForwardProp: (prop) => prop !== 'open',
+})(({ theme }) => ({
+    boxSizing: 'border-box',
+    flexShrink: 0,
+    variants: [
+        {
+            props: ({ open }) => open,
+            style: {
+                ...openedMixin(theme),
+                '& .MuiDrawer-paper': openedMixin(theme),
+            },
+        },
+        {
+            props: ({ open }) => !open,
+            style: {
+                ...closedMixin(theme),
+                '& .MuiDrawer-paper': closedMixin(theme),
+            },
+        },
+    ],
+    width: headerDrawerOpenWidth,
+    whiteSpace: 'nowrap',
+}));
