+++ src/components/SidebarStatic/SidebarStatic.tsx
@@ -2,7 +2,6 @@ import { FC } from 'react';
 import { Link } from 'react-router-dom';
 
 import {
-    Drawer as MuiDrawer,
     IconButton,
     List,
     ListItem,
@@ -10,9 +9,6 @@ import {
     ListItemIcon,
     ListItemText,
     useTheme,
-    styled,
-    Theme,
-    CSSObject,
 } from '@mui/material';
 import {
     ChevronLeft as IconChevronLeft,
@@ -22,62 +18,9 @@ import {
 import { navigation } from '../../constants/routerConstants';
 
 import type { IProps } from './SidebarStatic.types';
+import { Drawer, DrawerHeader } from './SidebarStatic.styles';
 
-const openWidth = 240;
-
-const openedMixin = (theme: Theme): CSSObject => ({
-    overflowX: 'hidden',
-    transition: theme.transitions.create('width', {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.enteringScreen,
-    }),
-    width: openWidth,
-});
-
-const closedMixin = (theme: Theme): CSSObject => ({
-    overflowX: 'hidden',
-    transition: theme.transitions.create('width', {
-        easing: theme.transitions.easing.sharp,
-        duration: theme.transitions.duration.leavingScreen,
-    }),
-    width: `calc(${theme.spacing(7)} + 1px)`,
-    [theme.breakpoints.up('sm')]: {
-        width: `calc(${theme.spacing(8)} + 1px)`,
-    },
-});
-
-const DrawerHeader = styled('div')(({ theme }) => ({
-    alignItems: 'center',
-    display: 'flex',
-    justifyContent: 'flex-end',
-    padding: theme.spacing(0, 1),
-    ...theme.mixins.toolbar,
-}));
-
-const Drawer = styled(MuiDrawer, {
-    shouldForwardProp: (prop) => prop !== 'open',
-})(({ theme }) => ({
-    boxSizing: 'border-box',
-    flexShrink: 0,
-    variants: [
-        {
-            props: ({ open }) => open,
-            style: {
-                ...openedMixin(theme),
-                '& .MuiDrawer-paper': openedMixin(theme),
-            },
-        },
-        {
-            props: ({ open }) => !open,
-            style: {
-                ...closedMixin(theme),
-                '& .MuiDrawer-paper': closedMixin(theme),
-            },
-        },
-    ],
-    width: openWidth,
-    whiteSpace: 'nowrap',
-}));
+const tempPermanentOpen = false;
 
 /**
  * Main navigation component, displays the page options.
@@ -98,21 +41,23 @@ const SidebarStatic: FC<IProps> = ({
     return (
         <Drawer onClose={onClose} open={open} variant='permanent'>
             <DrawerHeader>
-                <IconButton
-                    color='inherit'
-                    aria-label='open drawer'
-                    onClick={onToggle}
-                    edge='start'
-                    sx={{
-                        marginRight: 5,
-                    }}
-                >
-                    {theme.direction === 'rtl' ? (
-                        <IconChevronRight />
-                    ) : (
-                        <IconChevronLeft />
-                    )}
-                </IconButton>
+                {tempPermanentOpen ? null : (
+                    <IconButton
+                        color='inherit'
+                        aria-label='open drawer'
+                        onClick={onToggle}
+                        edge='start'
+                        sx={{
+                            marginRight: 5,
+                        }}
+                    >
+                        {theme.direction === 'rtl' ? (
+                            <IconChevronRight />
+                        ) : (
+                            <IconChevronLeft />
+                        )}
+                    </IconButton>
+                )}
             </DrawerHeader>
             <List>
                 {navigation.map(({ label, Icon, location }, index) => (
@@ -123,7 +68,9 @@ const SidebarStatic: FC<IProps> = ({
                     >
                         <Link to={location}>
                             <ListItemButton
-                                onClick={onClose}
+                                onClick={
+                                    tempPermanentOpen ? undefined : onClose
+                                }
                                 sx={{
                                     minHeight: 48,
                                     justifyContent: open ? 'initial' : 'center',
